==> 2- Ecommerce-Predicting_brand_col_RandomForest.ipynb <==
# **RandomForest to predict the 'brand' column's missing values**

## # Libraries Loading


import numpy     as np
import pandas    as pd

# Warnings
import warnings 
%matplotlib inline
warnings.filterwarnings('ignore')

# # Drive connect to get the data.
# from google.colab import drive
# drive.mount('/content/drive')


## # Data Loading

# # Import PyDrive and associated libraries.
# # This only needs to be done once per notebook.
# from pydrive.auth import GoogleAuth
# from pydrive.drive import GoogleDrive
# from google.colab import auth
# from oauth2client.client import GoogleCredentials

# # Authenticate and create the PyDrive client.
# # This only needs to be done once per notebook.
# auth.authenticate_user()
# gauth = GoogleAuth()
# gauth.credentials = GoogleCredentials.get_application_default()
# drive = GoogleDrive(gauth)

# # Download a file based on its file ID.
# #
# # A file ID looks like: laggVyWshwcyP6kEI-y_W3P8D26sz
# file_id = 'REPLACE_WITH_YOUR_FILE_ID'
# downloaded = drive.CreateFile({'id': file_id})
# print('Downloaded content "{}"'.format(downloaded.GetContentString()))

#data1 = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/Data 1.csv')
data1 = pd.read_csv('Data 1.csv')
data1

#data2 = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/Data 2.csv')
data2 = pd.read_csv('Data 2.csv')
data2

df1 = data1.copy()
df1

df2 = data2.copy()
df2

## Checking the RandomForest algorithm's performance on Predicting the 'brand' column's Missing values.

ds = pd.concat([df1, df2])

tiAll = ds.sample(n = 1000000)

### Checking for nulls and its percentage

# number of missing values and its percent
total = round(tiAll.isnull().sum(),2).sort_values(ascending=False)
percent = round((tiAll.isnull().sum()/tiAll.isnull().count()*100),2).sort_values(ascending=False)
missing_data = pd.concat([total, percent], axis=1, keys=['Total', 'Percent' + '%'])
missing_data

tiAll.loc[tiAll['brand'].isnull(),'brand'] = 'no brand'
tiAll

ti = tiAll[tiAll["brand"]!= 'no brand']
ti

### Checking for nulls and its percentage

# number of missing values and its percent
total = round(ti.isnull().sum(),2).sort_values(ascending=False)
percent = round((ti.isnull().sum()/ti.isnull().count()*100),2).sort_values(ascending=False)
missing_data = pd.concat([total, percent], axis=1, keys=['Total', 'Percent' + '%'])
missing_data

y = ti['brand']
y

#X = ti.drop('Survived', axis = 1)
X = ti[['product_id', 'category_id', 'price']]
X


### Splititng the data into trian and test data.

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

### RandomForest Algorithm accuracy %

# 5. Random Forest

from sklearn.ensemble import RandomForestClassifier
rfc = RandomForestClassifier()
rfc.fit(X_train, y_train)
y_pred = rfc.predict(X_test)
acc_rfc = round(rfc.score(X_train, y_train) * 100, 2)
print(acc_rfc,"%")



Since the Accuracy of the **RandomForest** Classifier algorithm in predicting the missing value in a **1M** rows data sample is **100%**, Then we can predict the missing values in the column 'brand' after spliting the data into parts around **1.5M** each to make it easier for the algorithm as for this accuracy there is no point of using the whole data in one time.

ds1 = data1.iloc[:1500000, :]
ds2 = data1.iloc[1500000 : 3000000, :]
ds3 = data1.iloc[3000000: , :]

print(ds1.shape, ds2.shape, ds3.shape)

ds4 = data2.iloc[:1500000, :]
ds5 = data2.iloc[1500000 : 3000000, :]
ds6 = data2.iloc[3000000: , :]


### ds1 Prediction accuracy

ds1.loc[ds1['brand'].isnull(),'brand'] = 'no brand'
ds11 = ds1[ds1["brand"]!= 'no brand']

y = ds11['brand']
X = ds11[['product_id', 'category_id', 'price']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

rfc = RandomForestClassifier()
rfc.fit(X_train, y_train)
y_pred = rfc.predict(X_test)
acc_rfc = round(rfc.score(X_train, y_train) * 100, 2)
print(acc_rfc,"%   accuracy for ds1")

### ds2 Prediction accuracy

ds2.loc[ds2['brand'].isnull(),'brand'] = 'no brand'
ds22 = ds2[ds2["brand"]!= 'no brand']

y = ds22['brand']
X = ds22[['product_id', 'category_id', 'price']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

rfc = RandomForestClassifier()
rfc.fit(X_train, y_train)
y_pred = rfc.predict(X_test)
acc_rfc = round(rfc.score(X_train, y_train) * 100, 2)
print(acc_rfc,"%   accuracy for ds2")

### ds3 Prediction accuracy

ds3.loc[ds3['brand'].isnull(),'brand'] = 'no brand'
ds33 = ds3[ds3["brand"]!= 'no brand']

y = ds33['brand']
X = ds33[['product_id', 'category_id', 'price']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

rfc = RandomForestClassifier()
rfc.fit(X_train, y_train)
y_pred = rfc.predict(X_test)
acc_rfc = round(rfc.score(X_train, y_train) * 100, 2)
print(acc_rfc,"%   accuracy for ds3")

### ds4 Prediction accuracy

ds4.loc[ds4['brand'].isnull(),'brand'] = 'no brand'
ds44 = ds4[ds4["brand"]!= 'no brand']

y = ds44['brand']
X = ds44[['product_id', 'category_id', 'price']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

rfc = RandomForestClassifier()
rfc.fit(X_train, y_train)
y_pred = rfc.predict(X_test)
acc_rfc = round(rfc.score(X_train, y_train) * 100, 2)
print(acc_rfc,"%   accuracy for ds4")

### ds5 Prediction accuracy

ds5.loc[ds5['brand'].isnull(),'brand'] = 'no brand'
ds55 = ds5[ds5["brand"]!= 'no brand']

y = ds55['brand']
X = ds55[['product_id', 'category_id', 'price']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

rfc = RandomForestClassifier()
rfc.fit(X_train, y_train)
y_pred = rfc.predict(X_test)
acc_rfc = round(rfc.score(X_train, y_train) * 100, 2)
print(acc_rfc,"%   accuracy for ds5")

### ds6 Prediction accuracy

ds6.loc[ds6['brand'].isnull(),'brand'] = 'no brand'
ds66 = ds6[ds6["brand"]!= 'no brand']

y = ds66['brand']
X = ds66[['product_id', 'category_id', 'price']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

rfc = RandomForestClassifier()
rfc.fit(X_train, y_train)
y_pred = rfc.predict(X_test)
acc_rfc = round(rfc.score(X_train, y_train) * 100, 2)
print(acc_rfc,"%   accuracy for ds6")

## Spliting the data into parts to perform RandomForest algorithm on it.

### Checking for nulls and its percentage

# number of missing values and its percent
total = round(ds1.isnull().sum(),2).sort_values(ascending=False)
percent = round((ds1.isnull().sum()/ds1.isnull().count()*100),2).sort_values(ascending=False)
missing_data = pd.concat([total, percent], axis=1, keys=['Total', 'Percent' + '%'])
missing_data

ds1.info()

## Predicting the Missing Values Using RandomForest Algorithm.

#### Creating a "no brand" category in the "brand" column 

ds1.loc[ds1['brand'].isnull(),'brand'] = 'no brand'

from sklearn.ensemble import RandomForestClassifier

brandNull = ds1[ds1["brand"] == 'no brand']
brandNotNull = ds1[ds1["brand"]!= 'no brand']
brandColumns = ['product_id', 'category_id', 'price']

rfModel_brand = RandomForestClassifier()

rfModel_brand.fit(brandNotNull[brandColumns], brandNotNull["brand"])
brandNullValues = rfModel_brand.predict(X= brandNull[brandColumns])

brandNull["brand"] = brandNullValues
ds1 = brandNotNull.append(brandNull)
ds1.reset_index(inplace=True)
ds1.drop('index', inplace=True, axis=1)


ds1

### Checking for nulls and its percentage

# number of missing values and its percent
total = round(ds1.isnull().sum(),2).sort_values(ascending=False)
percent = round((ds1.isnull().sum()/ds1.isnull().count()*100),2).sort_values(ascending=False)
missing_data = pd.concat([total, percent], axis=1, keys=['Total', 'Percent' + '%'])
missing_data

ds1.to_csv('ds1.csv', index = False)

---

ds2.loc[ds2['brand'].isnull(),'brand'] = 'no brand'

from sklearn.ensemble import RandomForestClassifier

brandNull = ds2[ds2["brand"] == 'no brand']
brandNotNull = ds2[ds2["brand"]!= 'no brand']
brandColumns = ['product_id', 'category_id', 'price']

rfModel_brand = RandomForestClassifier()

rfModel_brand.fit(brandNotNull[brandColumns], brandNotNull["brand"])
brandNullValues = rfModel_brand.predict(X= brandNull[brandColumns])

brandNull["brand"] = brandNullValues
ds2 = brandNotNull.append(brandNull)
ds2.reset_index(inplace=True)
ds2.drop('index', inplace=True, axis=1)


ds2.to_csv('ds2.csv', index = False)

---

ds3.loc[ds3['brand'].isnull(),'brand'] = 'no brand'

from sklearn.ensemble import RandomForestClassifier

brandNull = ds3[ds3["brand"] == 'no brand']
brandNotNull = ds3[ds3["brand"]!= 'no brand']
brandColumns = ['product_id', 'category_id', 'price']

rfModel_brand = RandomForestClassifier()

rfModel_brand.fit(brandNotNull[brandColumns], brandNotNull["brand"])
brandNullValues = rfModel_brand.predict(X= brandNull[brandColumns])

brandNull["brand"] = brandNullValues
ds3 = brandNotNull.append(brandNull)
ds3.reset_index(inplace=True)
ds3.drop('index', inplace=True, axis=1)


ds3.to_csv('ds3.csv', index = False)

---

ds4.loc[ds4['brand'].isnull(),'brand'] = 'no brand'

from sklearn.ensemble import RandomForestClassifier

brandNull = ds4[ds4["brand"] == 'no brand']
brandNotNull = ds4[ds4["brand"]!= 'no brand']
brandColumns = ['product_id', 'category_id', 'price']

rfModel_brand = RandomForestClassifier()

rfModel_brand.fit(brandNotNull[brandColumns], brandNotNull["brand"])
brandNullValues = rfModel_brand.predict(X= brandNull[brandColumns])

brandNull["brand"] = brandNullValues
ds4 = brandNotNull.append(brandNull)
ds4.reset_index(inplace=True)
ds4.drop('index', inplace=True, axis=1)


ds4.to_csv('ds4.csv', index = False)

---

ds5.loc[ds5['brand'].isnull(),'brand'] = 'no brand'

from sklearn.ensemble import RandomForestClassifier

brandNull = ds5[ds5["brand"] == 'no brand']
brandNotNull = ds5[ds5["brand"]!= 'no brand']
brandColumns = ['product_id', 'category_id', 'price']

rfModel_brand = RandomForestClassifier()

rfModel_brand.fit(brandNotNull[brandColumns], brandNotNull["brand"])
brandNullValues = rfModel_brand.predict(X= brandNull[brandColumns])

brandNull["brand"] = brandNullValues
ds5 = brandNotNull.append(brandNull)
ds5.reset_index(inplace=True)
ds5.drop('index', inplace=True, axis=1)


ds5.to_csv('ds5.csv', index = False)

---

ds6.loc[ds6['brand'].isnull(),'brand'] = 'no brand'

from sklearn.ensemble import RandomForestClassifier

brandNull = ds6[ds6["brand"] == 'no brand']
brandNotNull = ds6[ds6["brand"]!= 'no brand']
brandColumns = ['product_id', 'category_id', 'price']

rfModel_brand = RandomForestClassifier()

rfModel_brand.fit(brandNotNull[brandColumns], brandNotNull["brand"])
brandNullValues = rfModel_brand.predict(X= brandNull[brandColumns])

brandNull["brand"] = brandNullValues
ds6 = brandNotNull.append(brandNull)
ds6.reset_index(inplace=True)
ds6.drop('index', inplace=True, axis=1)


ds6.to_csv('ds6.csv', index = False)

---

##### Reading the CSV Files

d1 = pd.read_csv('ds1.csv')

d2 = pd.read_csv('ds2.csv')

d3 = pd.read_csv('ds3.csv')

d4 = pd.read_csv('ds4.csv')

d5 = pd.read_csv('ds5.csv')

d6 = pd.read_csv('ds6.csv')

#### Concating the 6 DataFrames to get the data in only ONE DataFrame

data = pd.concat([d1, d2, d3, d4, d5, d6])

data

data.shape

data.info()

### Checking for nulls and its percentage

# number of missing values and its percent
total = round(data.isnull().sum(),2).sort_values(ascending=False)
percent = round((data.isnull().sum()/data.isnull().count()*100),2).sort_values(ascending=False)
missing_data = pd.concat([total, percent], axis=1, keys=['Total', 'Percent' + '%'])
missing_data

## Exporting the Data after filling all the missing values in 'brand' column to a CSV file

data.to_csv('dataPredicted_RandomForest.csv', index = False)
