==> 3- Ecommerce-data_Preprocessing.ipynb <==
# !pip install chart_studio
# !pip install cufflinks
# !pip install plotly
# !pip install missingno


import numpy     as np
import pandas    as pd

# Options for pandas
pd.options.display.max_columns = 30

import matplotlib.pyplot as plt
import seaborn   as sns
import pylab
import calendar
from   scipy import stats as sci
import missingno as msno
from   datetime import datetime

# Standard plotly imports
import chart_studio.plotly as py
import plotly.graph_objs   as go
from   plotly.offline import iplot, init_notebook_mode

# Using plotly + cufflinks in offline mode
import cufflinks
cufflinks.go_offline(connected=True)
init_notebook_mode(connected=True)

# Set global theme
cufflinks.set_config_file(world_readable=True, theme="pearl")

# Display all cell outputs
from IPython.core.interactiveshell import InteractiveShell

InteractiveShell.ast_node_interactivity = "all"

import plotly.io as pio
pio.renderers.default = "colab"

# Warnings
import warnings 
%matplotlib inline
warnings.filterwarnings('ignore')

# Drive connect to get the data.
from google.colab import drive
drive.mount('/content/drive')



# Data Loading exported From RandomForest algorthm's Notebook
loading one data set for october sales and the other for november then combining Them


data_RF = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/dataPredicted_RandomForest.csv')
data_RF

data = data_RF.copy()


After concatenate data now we have 8738120 instance 

df = data.copy()

# DATA EXPLORATION & MANUPILATION

# df = data.copy().sample(n = 1000000)
df = data.copy()
# 'count_user' : how many times this user made an action on the site based on -partitioned by- the 'event_type'
df['count_user'] = df.groupby(['user_id', 'event_type'])['user_id'].transform('count')

# 'count_brand': how many times an action has been taken on this brand products based on the 'event_type'
df['count_brand'] = df.groupby(['brand', 'event_type'])['brand'].transform('count')

# 'count_event_type : how many times this event time has happened.
df['count_event_tye'] = df.groupby('event_type')['event_type'].transform('count')

# 'count_product' : how many times an action has been taken on this product based on the 'event_type'
df['count_product'] = df.groupby(['product_id', 'event_type'])['product_id'].transform('count')

# 'count-session' : Number of sessions do each user have.
df['count_session'] = df.groupby('user_session')['user_session'].transform('count')

# 'count_sessions_type' : Number of sessions do each user have per each 'event_type'
df['count_sessions_type'] = df.groupby(['user_session', 'event_type'])['product_id'].transform('count')

df

####Getting the size of the dataset.

df.shape 

###describe our data through some statistic methods

df.describe()

###know some more info about our data like data types & memory usage

df.info()

As we can see from above info() method, all the datatypes are correctly available. Hours is available as int64. Since hours is only used for grouping and other data analysis related works, its okay to have it in int64 format. Lets quickly check the minimum, average, maximum and other statics related to the numerical columns.

df['brand'].value_counts()

df['category_code'].value_counts()

# dropping dupplicates 
df.drop_duplicates()

## Feature Engineering for the 'event_time' colum

Event_time column has string "UTC" in it. So split the actual datetime and UTC seperate. Then convert the Event_time column into datetime datatype. Also, create new columns called date,time,hour,weekday,weeknum from the "event_time" column. Lets also change the weekday format from number to string like mon,tues,etc..

# 1- splitting the "Datetime" column into "date", "hour" , "year" columns
df["date"] = df.event_time.apply(lambda dateNum : dateNum.split()[0])
df["hour"] = df.event_time.apply(lambda hourNum : hourNum.split()[1].split(":")[0])
df["year"] = df.event_time.apply(lambda yearNum : yearNum.split()[0].split("-")[0])

# 2- Getting the "weekday", "month" columns out of "date" column
df["weekday"] = df.date.apply(lambda dateStr : calendar.day_name[datetime.strptime(dateStr,"%Y-%m-%d").weekday()])
df["month"] = df.date.apply(lambda dateStr : calendar.month_name[datetime.strptime(dateStr,"%Y-%m-%d").month])


### filling category code missings with other category

df.loc[df['category_code'].isnull(),'category_code'] = 'other'
df

### 'Event_type' Encoding Using ONE HOT ENCODING

event_type_enc = pd.get_dummies(df['event_type'])
event_type_enc

#### Adding the encoded columns into the original dataframe

df = pd.concat([df, event_type_enc], axis = 1)
df

### 'purchase' column ONE HOT ENCODING

purchase_enc = pd.get_dummies(df['purchase'])
purchase_enc

#### Renaming encoded columns

purchase_enc.rename(columns = {0:'no_purchase', 1 : 'yes_purchase'}, inplace = True)
purchase_enc

#### Adding purchase status columns into the df

df = pd.concat([df, purchase_enc], axis = 1)
df

### 'category_code' Encoding Using ONE HOT ENCODING

category_code_enc = pd.get_dummies(df['category_code'])
category_code_enc

#### Adding the encoded columns into the original dataframe

df = pd.concat([df, category_code_enc], axis = 1)
df

## Exporting the Manipulated data into file

now we have day, time , hour , year month and date

# df.to_csv('data_preprocessed.csv', index = False)
