==> 7-Ecommerce Recommendation System.ipynb <==
## Recommendation System Using SVD collapurative filtering 

Collaborative Filtering:-

This system matches persons with similar interests and provides recommendations based on this matching. Collaborative filters do not require item metadata like its content-based counterparts.
It is basically of two types:-



*   User based filtering:- 
These systems recommend products to a user that similar users have liked. For measuring the similarity between two users we can either use pearson correlation or cosine similarity.
*   Item Based Collaborative Filtering:-Instead of measuring the similarity between users, the item-based CF recommends items based on their similarity with the items that the target user rated. Likewise, the similarity can be computed with Pearson Correlation or Cosine Similarity. The major difference is that, with item-based collaborative filtering, we fill in the blank vertically, as oppose to the horizontal manner that user-based CF does. 


*  Single Value Decomposition:- One way to handle the scalability and sparsity issue created by CF is to leverage a latent factor model to capture the similarity between users and items. Essentially, we want to turn the recommendation problem into an optimization problem. We can view it as how good we are in predicting the rating for items given a user. One common metric is Root Mean Square Error (RMSE). The lower the RMSE, the better the performance.

!pip install scikit-surprise

# importing necessary library
import numpy as np
import pandas  as pd
from surprise import Reader, Dataset ,SVD
from surprise.model_selection import cross_validate

# Warnings
import warnings 
%matplotlib inline
warnings.filterwarnings('ignore')

# Drive connect to get the data.
from google.colab import drive
drive.mount('/content/drive')

df = pd.read_csv('/content/drive/MyDrive/Ecommerce Behavior Project/data_preprocessed.csv')

# collecting data that we will enter the model
matrix=df[df['event_type'] == 'purchase'].groupby(['user_id','product_id'],as_index=False).agg(no_of_purchases=('event_type','count')).sort_values(by='no_of_purchases',ascending=False)


matrix

reader= Reader()
df = Dataset.load_from_df(matrix[['user_id', 'product_id', 'no_of_purchases']], reader)


model = SVD()
cross_validate(model,df,measures=['rmse','mae'], cv=10)

We get a mean Root Mean Sqaure Error of 0.18 approx which is more than good enough for our case. Let us now train on our dataset and arrive at predictions.

let us remmemper:-

Lower RMSE values indicate better performance.

# Train the model using SVD
trainset =df.build_full_trainset()
model.fit(trainset)

let us pick user with user Id 401020145 and check the ratings she/he has given.


matrix[matrix['user_id'] == 401020145] 

# Generate predictions for the test set
model.predict(484176343,5751422,2)

If the SVD collaborative filter predicts a nuber_of_purchase is 1.01 for the product with ID 5788423 and user with ID 374211459, it means that the algorithm believes that this user who has not yet purchased this product is likely to purchase it around 1 times

The predicted rating is based on the patterns in the ratings data observed for other users and products. Specifically, the model has learned a set of latent factors that capture different aspects of user preferences and product characteristics. Based on the ratings patterns it has observed, the model has estimated the importance of each factor for predicting a rating for a given user and product. The predicted rating is then calculated by combining the user's preferences with the product's characteristics as represented by these latent factors.

It's important to note that the predicted rating is not necessarily an exact prediction of how this user will rate the product. It's an estimate based on patterns in the data, and there may be other factors that the model is not capturing. However, collaborative filtering algorithms like SVD have been shown to be effective at making accurate predictions in many cases, and can be useful for providing personalized recommendations to users.





---
